Add files via upload
Write a program that will output lines such as "===> <variable-name> = <value>" stating the value of named variables.


The program should read in variable assignment expressions one line at a time, and print out the value of variables as soon as 
they can be determined.  The input lines are given in the following format:


"<variable-name> = <arg> [ + <arg> ... ]" where <arg> is either a variable name or an integer.


Only addition expressions need to be supported, but any number of arguments can be added in the same expression.


Example execution, where lines starting with "===>" is written out by the program, and the rest of the lines are entered by the user:


a = 3
===> a = 3
b = 2 + 4
===> b = 6
c = a
===> c = 3
d = b + 7
===> d = 13
A = b + c + d
===> A = 22
f = g
g = 1 + 2 + 3
===> g = 6
===> f = 6
h = i + A + lemon
lemon = k + lima
i = A + 5
===> i = 27
k = a + b
===> k = 9
lima = k + k
===> lima = 18
===> lemon = 27
===> h = 76


Pay close attention to the cases where expressions use variables that hasn't been assigned yet as arguments.


Note that the application should print out the value of a variable as soon as it can be determined, rather than wait
until the user has entered all the assignment expressions.


The program does not need to behave in any particular way if invalid expressions are entered.
The program does not need to have any mechanism for ending the program.
The code for entering and printing out data is not important.  The example execution above suggests console based input, but any
user interface providing the same functionality is acceptable.
You may choose to skip the input gathering code all together and simulate input by defining and using functions such as
"void parse_line(string line)" or even "void add_assignment_expression(string target_variable_name, string[] sum_operation_args)".


The code will be evaluated according to the following criteria:


* That the code works correctly and performs the task described in this text.


* That the code is easily maintainable, i.e. easy to read and understand, and can easily be
   adopted to use in a large program.


* That the implementation is efficient and scales well when given large amounts of input data.